updated products route to account for product styles with no affiliated product
