+++ package.json
@@ -6,7 +6,10 @@
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "test": "jest --detectOpenHandles"
+    "test": "jest --detectOpenHandles",
+    "newRelic": "node -r newrelic server.js",
+    "k6": "k6 run k6/load-test.js",
+    "k6-last": "k6 run k6/last-test.js"
   },
   "repository": {
     "type": "git",
